# contract/openapi.yaml
openapi: 3.1.0
info:
  title: PaytmFlow API
  version: "1.0.0"
  description: >
    Canonical API contract for PaytmFlow (spec v4.2).
    FROZEN at T+3. Changes require the contract-change protocol in SHARED_CONTRACT.md §9.
    Dev2 owns this file. Dev1 generates TypeScript types from it and must never hand-edit them.
servers:
  - url: http://localhost:8000
    description: Local
  - url: /
    description: Deployed (same origin via proxy)
...
paths:
...
  /api/v1/translate:
    post:
      tags: [ai]
      operationId: translateText
      description: >
        Translates ONE already-generated canonical explanation/chat string
        (Sarvam Mayura). Never applied to amounts, statuses, display_value,
        or resume_screen - the frontend calls this per free-text block, never
        on the whole API response.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TranslateRequest" }
      responses:
        "200":
          description: Translated text
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TranslateResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
        "504":
          description: Translation timed out
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  /api/v1/review/role:
    post:
      tags: [review]
      operationId: switchPrototypeRole
      description: >
        Prototype-only: flips the CURRENT session's own role between CUSTOMER
        and REVIEW_OFFICER. This is not enterprise auth - see the Review
        Center product notes. Every /review/* endpoint still independently
        enforces the resulting role server-side.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { $ref: "#/components/schemas/ReviewerRole" }
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                type: object
                required: [role]
                properties:
                  role: { $ref: "#/components/schemas/ReviewerRole" }
  /api/v1/review/dashboard:
    get:
      tags: [review]
      operationId: getReviewDashboardMetrics
      description: Real, server-computed queue counts - never hard-coded.
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewDashboardMetrics" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/review/cases:
    get:
      tags: [review]
      operationId: listReviewQueue
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/ReviewCaseStatus" } }
        - { name: priority, in: query, schema: { $ref: "#/components/schemas/ReviewCasePriority" } }
        - { name: journey_type, in: query, schema: { $ref: "#/components/schemas/JourneyType" } }
        - { name: mine, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Review queue
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewQueueResponse" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/review/cases/{case_id}:
    get:
      tags: [review]
      operationId: getReviewCase
      parameters: [{ $ref: "#/components/parameters/CaseIdPath" }]
      responses:
        "200":
          description: Case detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewCaseDetail" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/review/cases/{case_id}/customer-view:
    get:
      tags: [review]
      operationId: getReviewCaseCustomerView
      description: Read-only - lets a reviewer see exactly what the customer currently sees.
      parameters: [{ $ref: "#/components/parameters/CaseIdPath" }]
      responses:
        "200":
          description: Case detail, as the customer sees their journey
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewCaseDetail" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/review/cases/{case_id}/audit:
    get:
      tags: [review]
      operationId: getReviewCaseAudit
      parameters: [{ $ref: "#/components/parameters/CaseIdPath" }]
      responses:
        "200":
          description: Full audit timeline for this case's journey
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewCaseAuditResponse" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/review/cases/{case_id}/ai-summary:
    post:
      tags: [review]
      operationId: summarizeReviewCase
      description: >
        Advisory-only "AI Evidence Summary", built from the same structured
        case data already shown to the reviewer. Never a decision -
        deterministic case-state rules remain authoritative.
      parameters: [{ $ref: "#/components/parameters/CaseIdPath" }]
      responses:
        "200":
          description: AI-generated advisory summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewSummaryResponse" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/review/cases/{case_id}/claim:
    post:
      tags: [review]
      operationId: claimReviewCase
      parameters: [{ $ref: "#/components/parameters/CaseIdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expected_case_version]
              properties:
                expected_case_version: { type: integer }
      responses:
        "200":
          description: Case claimed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewCase" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/ReviewCaseConflict" }
  /api/v1/review/cases/{case_id}/resolve:
    post:
      tags: [review]
      operationId: resolveReviewCase
      description: >
        Never approves/rejects a financial product - resolves the EVIDENCE
        exception. EVIDENCE_SUFFICIENT re-enters the exact same deterministic
        mutation chain a customer's own clarification answer uses.
      parameters: [{ $ref: "#/components/parameters/CaseIdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution_type, resolution_reason, expected_case_version]
              properties:
                resolution_type: { $ref: "#/components/schemas/ReviewResolutionType" }
                resolution_reason: { type: string, minLength: 1 }
                resolution_notes: { type: string }
                resolution_value: {}
                expected_case_version: { type: integer }
      responses:
        "200":
          description: Case resolved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewCaseDetail" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/ReviewCaseConflict" }
  /api/v1/review/cases/{case_id}/request-information:
    post:
      tags: [review]
      operationId: requestReviewCaseInformation
      parameters: [{ $ref: "#/components/parameters/CaseIdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requested_docs, customer_message, expected_case_version]
              properties:
                requested_docs: { type: array, items: { type: string }, minItems: 1 }
                customer_message: { type: string, minLength: 1 }
                expected_case_version: { type: integer }
      responses:
        "200":
          description: Information requested; case moves to ADDITIONAL_INFO_REQUIRED
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewCase" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/ReviewCaseConflict" }
  /api/v1/review/cases/{case_id}/escalate:
    post:
      tags: [review]
      operationId: escalateReviewCase
      parameters: [{ $ref: "#/components/parameters/CaseIdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [escalation_reason, expected_case_version]
              properties:
                escalation_reason: { type: string, minLength: 1 }
                expected_case_version: { type: integer }
      responses:
        "200":
          description: Case escalated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReviewCase" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/ReviewCaseConflict" }
  /api/v1/journeys/{journey_id}/review-status:
    get:
      tags: [review]
      operationId: getJourneyReviewStatus
      description: >
        Customer-facing, customer-safe phrasing only - never exposes internal
        case metadata, reviewer identity, or AI confidence numbers.
      parameters: [{ $ref: "#/components/parameters/JourneyIdPath" }]
      responses:
        "200":
          description: Review status for this journey
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerReviewStatus" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pf_session
    demoSecret:
      type: apiKey
      in: header
      name: X-Demo-Secret
  parameters:
    JourneyIdPath:
      name: journey_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CaseIdPath:
      name: case_id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    NotFound:
      description: Not found, or not owned by this session (deliberately indistinguishable)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    ValidationError:
      description: VALIDATION_ERROR or INVALID_JOURNEY_TYPE
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    Forbidden:
      description: FORBIDDEN. The current session's role does not permit this action.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    ReviewCaseConflict:
      description: >
        REVIEW_CASE_STALE (case_version mismatch), REVIEW_CASE_LOCKED (claimed by
        another reviewer), or REVIEW_CASE_INVALID_TRANSITION. No mutation occurred.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    JourneyType:
      type: string
      enum: [LENDING, INSURANCE, CREDIT_CARD, KYC, ACCOUNT_OPENING, INVESTMENT]
    Readiness:
      type: string
      enum: [READY, NOT_READY, NEEDS_REVIEW, DEAD_END]
      description: Deterministic enum. There is no readiness score anywhere in this API.
    JourneyStatus:
      type: string
      enum: [IN_PROGRESS, COMPLETED, NEEDS_REVIEW]
    FieldStatus:
      type: string
      enum: [SATISFIED, BLOCKED, AMBIGUOUS]
    ErrorCode:
      type: string
      enum:
        - INVALID_JOURNEY_TYPE
        - VALIDATION_ERROR
        - ACTION_INVALID
        - ACTION_STALE
        - NEEDS_REVIEW
        - DEAD_END
        - AI_TIMEOUT
        - AI_MALFORMED_OUTPUT
        - EVIDENCE_CONFLICT
        - NOT_FOUND
        - FORBIDDEN
        - REVIEW_CASE_STALE
        - REVIEW_CASE_LOCKED
        - REVIEW_CASE_INVALID_TRANSITION
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message:
              type: string
              description: Human-readable and safe to render directly to the user.
            details:
              type: object
              additionalProperties: true
            current_snapshot_id:
              type: string
              format: uuid
              description: Present on ACTION_STALE so the client can resync without a second call.
...
    FieldState:
      type: object
      required: [key, label, status]
      properties:
        key: { type: string }
        label: { type: string, example: "Income verification" }
        status: { $ref: "#/components/schemas/FieldStatus" }
        value: {}
        display_value: { type: string, description: Pre-formatted for display, e.g. "₹85,000" }
        explanation:
          type: string
          description: Why this is blocked. Rendered on the blocker card.
        resolve_action_id:
          type: string
          description: The action that resolves this field. Powers the "Resolve ->" button.
        mandatory: { type: boolean }
        ambiguity:
          type: object
          nullable: true
          properties:
            ambiguity_id: { type: string }
            reason: { type: string }
            question: { type: string }
            answer_type: { type: string, enum: [CHOICE, MONEY, NUMBER, TEXT, DATE, BOOLEAN] }
            choices:
              type: array
              items:
                type: object
                properties:
                  value: {}
                  label: { type: string }
    ProgressCounts:
      type: object
      required: [completed, pending, blockers, total]
      description: >
        Journey progress only. NOT readiness, approval, eligibility or a score.
        Dev1 renders this as "3/7 Completed". Never render it as a percentage or gauge.
      properties:
        completed: { type: integer }
        pending: { type: integer }
        blockers: { type: integer }
        total: { type: integer }
    JourneyStateResponse:
      type: object
      required: [journey_id, journey_type, schema_version, snapshot_id, version_number,
                 readiness, status, fields, progress, display]
      properties:
        journey_id: { type: string, format: uuid }
        journey_type: { $ref: "#/components/schemas/JourneyType" }
        schema_version: { type: string }
        snapshot_id: { type: string, format: uuid }
        version_number: { type: integer }
        readiness: { $ref: "#/components/schemas/Readiness" }
        status: { $ref: "#/components/schemas/JourneyStatus" }
        fields:
          type: array
          items: { $ref: "#/components/schemas/FieldState" }
          description: Ordered for display. Only fields with display=true and not NOT_APPLICABLE.
        progress: { $ref: "#/components/schemas/ProgressCounts" }
        pending_clarification:
          allOf: [{ $ref: "#/components/schemas/FieldState" }]
          nullable: true
          description: Present when readiness is NEEDS_REVIEW. Exactly one at a time.
        display:
          type: object
          required: [title, summary]
          properties:
            title: { type: string, example: "Personal Loan" }
            summary: { type: string, example: "₹5,00,000 · Home Renovation" }
        updated_at: { type: string, format: date-time }
...
    JourneyDiff:
      type: object
      required: [from_version, to_version, fields_changed, actions_unlocked, actions_removed, readiness, progress]
      properties:
        from_version: { type: integer }
        to_version: { type: integer }
        fields_changed:
          type: array
          items:
            type: object
            required: [key, label, from_status, to_status]
            properties:
              key: { type: string }
              label: { type: string }
              from_status: { $ref: "#/components/schemas/FieldStatus" }
              to_status: { $ref: "#/components/schemas/FieldStatus" }
              display_value: { type: string }
              cause: { type: string, example: "ACTION:UPLOAD_INCOME_PROOF" }
              cascaded: { type: boolean, description: true if changed by cascade, not directly by the action }
        actions_unlocked: { type: array, items: { type: string } }
        actions_removed: { type: array, items: { type: string } }
        readiness:
          type: object
          properties:
            from: { $ref: "#/components/schemas/Readiness" }
            to: { $ref: "#/components/schemas/Readiness" }
        progress:
          type: object
          properties:
            from: { $ref: "#/components/schemas/ProgressCounts" }
            to: { $ref: "#/components/schemas/ProgressCounts" }
...
    TranslateRequest:
      type: object
      required: [text, target_language_code]
      properties:
        text: { type: string, minLength: 1, maxLength: 2000 }
        target_language_code: { type: string, minLength: 2, maxLength: 10 }
        source_language_code: { type: string, default: auto }
    TranslateResponse:
      type: object
      required: [translated_text]
      properties:
        translated_text: { type: string }
...
    ReviewerRole:
      type: string
      enum: [CUSTOMER, REVIEW_OFFICER]
    ReviewCaseStatus:
      type: string
      enum:
        - REVIEW_REQUIRED
        - UNDER_REVIEW
        - ADDITIONAL_INFO_REQUIRED
        - RESOLVED
        - ESCALATED
        - CANCELLED
    ReviewCasePriority:
      type: string
      enum: [LOW, MEDIUM, HIGH]
    ReviewResolutionType:
      type: string
      enum:
        - EVIDENCE_SUFFICIENT
        - REQUEST_ADDITIONAL_INFORMATION
        - EVIDENCE_CANNOT_BE_VERIFIED
        - CUSTOMER_INFO_NEEDS_CORRECTION
        - ESCALATE_FOR_SPECIALIST_REVIEW
    ReviewEvidenceSummary:
      type: object
      required: [evidence_id, doc_type, filename, uploaded_at, verified]
      properties:
        evidence_id: { type: string, format: uuid }
        doc_type: { type: string }
        filename: { type: string }
        uploaded_at: { type: string, format: date-time }
        verified: { type: boolean }
        confidence: { type: number }
        extracted_values: { type: object, additionalProperties: true }
        provider:
          type: string
          nullable: true
          description: >
            Which AIProvider produced this extraction ("sarvam", "local_ml",
            "llm", "mock") - source traceability for the Review Center.
    ReviewSummaryResponse:
      type: object
      required: [summary, disclaimer]
      properties:
        summary: { type: string }
        disclaimer: { type: string }
    ReviewCase:
      type: object
      required:
        - case_id
        - case_number
        - journey_id
        - journey_type
        - journey_display_name
        - field_key
        - reason_code
        - reason_title
        - reason_description
        - priority
        - status
        - case_version
        - created_at
        - updated_at
      properties:
        case_id: { type: string, format: uuid }
        case_number: { type: string, example: "PF-0001" }
        journey_id: { type: string, format: uuid }
        journey_type: { $ref: "#/components/schemas/JourneyType" }
        journey_display_name: { type: string }
        field_key: { type: string }
        reason_code: { type: string }
        reason_title: { type: string }
        reason_description: { type: string }
        priority: { $ref: "#/components/schemas/ReviewCasePriority" }
        status: { $ref: "#/components/schemas/ReviewCaseStatus" }
        case_version: { type: integer }
        assigned_reviewer: { type: string, nullable: true }
        is_locked: { type: boolean }
        resolution_type: { $ref: "#/components/schemas/ReviewResolutionType" }
        resolution_reason: { type: string, nullable: true }
        resolution_notes: { type: string, nullable: true }
        requested_information:
          type: object
          nullable: true
          properties:
            requested_docs: { type: array, items: { type: string } }
            customer_message: { type: string }
        escalation_reason: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        resolved_at: { type: string, format: date-time, nullable: true }
    ReviewCaseDetail:
      allOf:
        - $ref: "#/components/schemas/ReviewCase"
        - type: object
          required: [journey_context, evidence]
          properties:
            journey_context: { $ref: "#/components/schemas/JourneyStateResponse" }
            evidence:
              type: array
              items: { $ref: "#/components/schemas/ReviewEvidenceSummary" }
            impact_preview:
              $ref: "#/components/schemas/JourneyDiff"
              description: >
                Deterministic, non-persisting preview of resolving this case
                EVIDENCE_SUFFICIENT - null when no evidence-derived value
                exists yet to preview.
    ReviewQueueResponse:
      type: object
      required: [cases]
      properties:
        cases:
          type: array
          items: { $ref: "#/components/schemas/ReviewCase" }
    ReviewDashboardMetrics:
      type: object
      required: [pending_review, under_review, waiting_customer, escalated, resolved_today]
      properties:
        pending_review: { type: integer }
        under_review: { type: integer }
        waiting_customer: { type: integer }
        escalated: { type: integer }
        resolved_today: { type: integer }
    ReviewCaseAuditEntry:
      type: object
      required: [event_type, payload, created_at]
      properties:
        event_type: { type: string }
        payload: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
    ReviewCaseAuditResponse:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items: { $ref: "#/components/schemas/ReviewCaseAuditEntry" }
    CustomerReviewStatus:
      type: object
      required: [has_open_case]
      properties:
        has_open_case: { type: boolean }
        case_number: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/ReviewCaseStatus" }
        title: { type: string, nullable: true }
        description: { type: string, nullable: true }
        requested_information:
          type: object
          nullable: true
          properties:
            requested_docs: { type: array, items: { type: string } }
            customer_message: { type: string }
        updated_at: { type: string, format: date-time, nullable: true }
security:
  - sessionCookie: []